* Fix the size of the filler of the progress bar in the pulse mode * Don't expand the table when the table had no expandable children!
git-svn-id: http://svn.enlightenment.org/svn/e/trunk@23555 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33

+++ e17/libs/etk/src/lib/etk_progress_bar.c
@@ -321,7 +321,7 @@ static void _etk_progress_bar_property_get(Etk_Object *object, int property_id, 
 static void _etk_progress_bar_realize_cb(Etk_Object *object, void *data)
 {
    Etk_Progress_Bar *progress_bar;
-   double w;
+   float w;
    
    if (!(progress_bar = ETK_PROGRESS_BAR(object)))
       return;
